Split out all renderers
